requirements.txt
Flask==2.3.3
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
openai==1.3.0
sentence-transformers==2.2.2
pypdf==3.17.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
python-docx==1.1.0
PyPDF2==3.0.1
